Restructure converter specs
